habit_tracker_agent/requirements.txt
google-generativeai>=0.8.0
gspread>=6.0.0
google-auth>=2.25.0
requests>=2.31.0
pytz>=2024.1
functions-framework>=3.5.0
